Organize directory of handlers
