+++ functions/handlers/fetch_booking_list.py
@@ -1,7 +1,7 @@
 import os
 import sys
 
-FUNCTIONS_DIR_PATH = os.path.dirname(__file__)
+FUNCTIONS_DIR_PATH = os.path.dirname(os.path.dirname(__file__))
 sys.path.append(FUNCTIONS_DIR_PATH)
 
 from domain.booking import BookingTable
