+++ functions/hub.py
@@ -1,0 +1,1 @@
+from fastapi import FastAPI
